parity: the hire upsert and the research read cross — a SQL comment the line-continuations de-newlined, and an int4 decoded as int8
agent_defs/gaps: TS's queries carried -- comments; ported into \\-continued
string literals the newlines vanish, so the comment swallowed the rest of
the statement (role coalesce, returning — syntax error at end of input,
every hire and every gap record). Comments moved out of the SQL.

research: research_sources.idx is int4; both readers (route + engine
sources_of) decoded it as i64 — 500 on any run with sources, which is why
the battery never saw it. Now i32 both places.

+++ api/src/research.rs
@@ -561,7 +561,7 @@ pub async fn get_research_run(
         .await?;
     let Some(row) = row else { return Ok(None) };
     let run = row_of(&row);
-    let sources: Vec<(i64, String, Option<String>, Option<String>)> = sqlx::query_as(
+    let sources: Vec<(i32, String, Option<String>, Option<String>)> = sqlx::query_as(
         "select idx, url, title, snippet from research_sources \
          where run_id = $1::uuid order by idx asc",
     )
